T3

+++ lib/view/home/components/my_group_part.dart
@@ -1,14 +1,20 @@
 import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:hide_out/%20data_models/group.dart';
+import 'package:hide_out/%20data_models/notification.dart' as d;
+import 'package:hide_out/utils/constants.dart';
+import 'package:hide_out/utils/style.dart';
+import 'package:hide_out/view/common/items/dialog/custom_alert_dialog.dart';
+import 'package:hide_out/view/common/items/dialog/help_dialog.dart';
+import 'package:hide_out/view/group/group_screen.dart';
+import 'package:hide_out/view_models/home_screen_view_model.dart';
 import 'package:provider/provider.dart';
-import 'package:voice_put/%20data_models/group.dart';
-import 'package:voice_put/%20data_models/notification.dart' as d;
-import 'package:voice_put/utils/constants.dart';
-import 'package:voice_put/view/common/items/dialog/help_dialog.dart';
-import 'package:voice_put/view/group/group_screen.dart';
-import 'package:voice_put/utils/style.dart';
-import 'package:voice_put/view_models/home_screen_view_model.dart';
 
 class MyGroupPart extends StatelessWidget {
+  MyGroupPart({Key? key, required this.globalKey}) : super(key: key);
+
+  final GlobalKey globalKey;
+
   @override
   Widget build(BuildContext context) {
     final homeScreenViewModel = context.read<HomeScreenViewModel>();
@@ -23,6 +29,7 @@ class MyGroupPart extends StatelessWidget {
           padding: const EdgeInsets.only(left: 24.0),
           child: Text(
             "My Group",
+            key: globalKey,
             style: homeScreenLabelTextStyle,
           ),
         ),
@@ -36,8 +43,10 @@ class MyGroupPart extends StatelessWidget {
             } else {
               if (model.isFirstCall) {
                 homeScreenViewModel.stopCall();
+                _showAlertAutoExitDialog(context, model.notifications);
                 _showAutoExitDialog(context, model.notifications);
                 _showDeletedGroupDialog(context, model.notifications);
+                _showNewOwnerDialog(context, model.notifications);
               }
 
               return model.groups.isEmpty
@@ -82,6 +91,36 @@ class MyGroupPart extends StatelessWidget {
     );
   }
 
+  void _showAlertAutoExitDialog(
+      BuildContext context, List<d.Notification> notifications) async {
+    final homeScreenViewModel = context.read<HomeScreenViewModel>();
+
+    var alertNotifications = notifications.where((element) =>
+        element.notificationType == NotificationType.ALERT_AUTO_EXIT);
+
+    if (alertNotifications.isNotEmpty) {
+      alertNotifications.forEach((element) {
+        Future(() => showCustomAlertDialog(
+            context: context,
+            titleStr: "Warning!",
+            contentStr:
+                'You are about to be kicked out of "${element.content}". Please make a recording TODAY!',
+            confirmedStr: "I will do it!",
+            circleColor: Colors.red.shade400,
+            icon: Icon(
+              FontAwesomeIcons.calendarAlt,
+              size: 28.0,
+              color: Colors.white,
+            ),
+            onConfirmed: () {
+              homeScreenViewModel.deleteNotification(element.notificationId);
+              homeScreenViewModel.updateIsAlerted(
+                  element.groupId!, element.userId!);
+            }));
+      });
+    }
+  }
+
   void _showAutoExitDialog(
       BuildContext context, List<d.Notification> notifications) async {
     final homeScreenViewModel = context.read<HomeScreenViewModel>();
@@ -91,13 +130,20 @@ class MyGroupPart extends StatelessWidget {
 
     if (autoExitNotifications.isNotEmpty) {
       autoExitNotifications.forEach((element) {
-        Future(() => showHelpDialog(
+        Future(() => showCustomAlertDialog(
             context: context,
-            title: Text("Auto-Exit Period"),
-            contentString:
-                'You have exited from "${element.content}". Please enter again if you want.',
-            okayString: "Okay",
+            titleStr: "Sorry to see you go!",
+            contentStr: "You were kicked out of '${element.content}'. "
+                "But don't cry! You can join the group again!!",
+            confirmedStr: "Okay!",
+            circleColor: Colors.blue.shade400,
+            icon: Icon(
+              FontAwesomeIcons.sadTear,
+              color: Colors.white,
+              size: 28.0,
+            ),
             onConfirmed: () {
+              print('pushed');
               homeScreenViewModel.deleteNotification(element.notificationId);
             }));
       });
@@ -126,6 +172,28 @@ class MyGroupPart extends StatelessWidget {
     }
   }
 
+  void _showNewOwnerDialog(
+      BuildContext context, List<d.Notification> notifications) {
+    final homeScreenViewModel = context.read<HomeScreenViewModel>();
+
+    var newOwnerNotifications = notifications.where(
+        (element) => element.notificationType == NotificationType.NEW_OWNER);
+
+    if (newOwnerNotifications.isNotEmpty) {
+      newOwnerNotifications.forEach((element) {
+        Future(() => showHelpDialog(
+            context: context,
+            title: Text("You are the owner!"),
+            contentString:
+                'In "${element.content}", you have become the new owner because the previous owner exited from the group.',
+            okayString: "Okay",
+            onConfirmed: () {
+              homeScreenViewModel.deleteNotification(element.notificationId);
+            }));
+      });
+    }
+  }
+
   Widget _myGroupListView(
       List<Group> groups, List<d.Notification> notifications) {
     return ListView.builder(
@@ -137,7 +205,7 @@ class MyGroupPart extends StatelessWidget {
             padding: const EdgeInsets.symmetric(horizontal: 16.0),
             child: Stack(alignment: Alignment.topRight, children: [
               Card(
-                color: listTileColor,
+                color: darkBackgroundButtonColor,
                 shape: RoundedRectangleBorder(
                   borderRadius: BorderRadius.circular(12.0),
                 ),
@@ -150,7 +218,7 @@ class MyGroupPart extends StatelessWidget {
                       padding: const EdgeInsets.symmetric(vertical: 18.0),
                       child: Text(
                         group.groupName!,
-                        style: listTileTitleTextStyle,
+                        style: darkBackgroundListTileTextStyle,
                       ),
                     ),
                     trailing: Icon(Icons.arrow_forward_ios_rounded),
@@ -170,9 +238,7 @@ class MyGroupPart extends StatelessWidget {
                       ),
                       child: Center(
                         child: Text(
-                            "${notifications.where((element)
-                            => element.notificationType == NotificationType.NEW_POST
-                               && element.groupId == group.groupId).length}"),
+                            "${notifications.where((element) => element.notificationType == NotificationType.NEW_POST && element.groupId == group.groupId).length}"),
                       ),
                     )
                   : Container(),
